Add: getProcessedDiaryList 추가
