+++ src/components/organisms/item/MemoListItem.js
@@ -1,39 +1,16 @@
 import styled from "styled-components";
 import { useState } from "react";
 import { useNavigate } from "react-router-dom";
+import { sortOptionList, filterOptionList } from "../../../util/optionList";
+import { getProcessedDiaryList } from "../../../util/optionList";
 import Button from "../../molecule/etc/Button";
 import MemoItemBox from "../box/MemoItemBox";
 import SelectBox from "../box/SelectBox";
-import { sortOptionList, filterOptionList } from "../../../util/optionList";
 
 const MemoListItem = ({ diaryList }) => {
   const navigate = useNavigate();
   const [sortType, setSortType] = useState("latest");
   const [filter, setFilter] = useState("all");
-
-  const getProcessedDiaryList = () => {
-    const filterCallBack = (item) => {
-      if (filter === "good") {
-        return parseInt(item.emotion) <= 3;
-      } else {
-        return parseInt(item.emotion) > 3;
-      }
-    };
-    const compare = (a, b) => {
-      if (sortType === "latest") {
-        return parseInt(b.date) - parseInt(a.date);
-      } else {
-        return parseInt(a.date) - parseInt(b.date);
-      }
-    };
-    const copyList = JSON.parse(JSON.stringify(diaryList));
-    let filterList =
-      filter === "all" ? copyList : copyList.filter((it) => filterCallBack(it));
-    const sortedList = filterList.sort(compare);
-
-    return sortedList;
-  };
-
   const goNew = () => {
     navigate("/new");
   };
@@ -60,8 +37,9 @@ const MemoListItem = ({ diaryList }) => {
           />
         </RightCol>
       </Wrapper>
-      {diaryList &&
-        diaryList.map((list) => <MemoItemBox key={list.id} {...list} />)}
+      {getProcessedDiaryList(filter, sortType, diaryList).map((list) => (
+        <MemoItemBox key={list.id} {...list} />
+      ))}
     </MemoList>
   );
 };
